finished product data modal

+++ static/js/script.js
@@ -6,19 +6,23 @@ const newCategoryInput = document.getElementById('newCategoryInput');
 //Categories page ------------------------------------------------------------
 
 const categoryTable = document.getElementById('category-table')
-const categoryButtons = categoryTable.querySelectorAll('#delete-category-btn').forEach(btn => {
-    btn.addEventListener('click', function () {
-        const id = this.getAttribute('data-id')
+if (categoryTable) {
+    categoryTable.querySelectorAll('#delete-category-btn').forEach(btn => {
+        btn.addEventListener('click', function () {
+            const id = this.getAttribute('data-id')
 
-        if (confirm('Tem certeza que deseja remover esta categoria?')) {
-            fetch(`/delete_category/${id}`, {
-                method: 'DELETE',
-                header: headers
-            }).then(() => window.location.reload())
-        }
+            if (confirm('Tem certeza que deseja remover esta categoria?')) {
+                fetch(`/delete_category/${id}`, {
+                    method: 'DELETE',
+                    header: headers
+                }).then(() => window.location.reload())
+            }
 
+        })
     })
-})
+}
+
+
 //New category page ----------------------------------------------------------
 
 newCategoryForm.addEventListener('submit', (event) => {
